pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = ""
authors = ["Tomas Pytel <pytlicek@gmail.com>"]
readme = "README.md"
packages = [{include = "django_template"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
black = "^23.7.0"
django-ninja = "^0.22.2"
django-environ = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
